backend/requirements.txt
fastapi~=0.115.11
uvicorn
PyPDF2
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
openai
langchain-openai
python-dotenv
python-multipart
pydantic
tiktoken
faiss-cpu>=1.8.1
